pyproject.toml
[project]
name = "nyc-taxi-duration-mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "catboost>=1.2.8",
    "fastparquet>=2025.12.0",
    "flask>=3.1.2",
    "hyperopt>=0.2.7",
    "mlflow>=3.8.1",
    "pandas>=2.3.3",
    "pandas-stubs>=2.3.3.251219",
    "pyarrow>=22.0.0",
    "scikit-learn>=1.8.0",
    "xgboost>=3.1.2",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "hypothesis>=6.148.11",
    "pre-commit>=4.5.1",
    "pyright>=1.1.407",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]


[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
# This ensures that even with the 'src' layout, pytest finds your code
addopts = "-ra -q"